requirements.txt
openai
python-dotenv
streamlit
PyPDF2
azure-search-documents
AzureOpenAI
numpy
operator